+++ unittest/double-entry-container.cpp
@@ -119,7 +119,7 @@ BOOST_AUTO_TEST_CASE(test_all_matrix_stack)
 {
   typedef Eigen::Matrix<double, 6, 6> Matrix6;
 
-  typedef MatrixStackTpl<Matrix6> Vector;
+  typedef internal::MatrixStackTpl<Matrix6> Vector;
   typedef container::DoubleEntryContainer<Vector> Container;
 
   const Eigen::Index nrows = 20, ncols = 20;